mnout/src/main/webapp/app/entities/lob/delete/lob-delete-dialog.component.ts
import { Component } from '@angular/core';
import { FormsModule } from '@angular/forms';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import SharedModule from 'app/shared/shared.module';
import { ILOB } from '../lob.model';
import { LOBService } from '../service/lob.service';
import { ITEM_DELETED_EVENT } from 'app/config/navigation.constants';

@Component({
  standalone: true,
  templateUrl: './lob-delete-dialog.component.html',
  imports: [SharedModule, FormsModule],
})
export class LOBDeleteDialogComponent {
  lOB?: ILOB;

  constructor(protected lOBService: LOBService, protected activeModal: NgbActiveModal) {}

  cancel(): void {
    this.activeModal.dismiss();
  }

  confirmDelete(id: number): void {
    this.lOBService.delete(id).subscribe(() => {
      this.activeModal.close(ITEM_DELETED_EVENT);
    });
  }
}
